NPT-1038: Migrate Custom Attribute editor to routed page + fix delete cascade
Round 3 rework on the Custom Attribute Types admin (#1038).

UI migration:
- Replaces the create/edit modal with a routed page at
  /manage/custom-attributes/new and /:id/edit, mirroring the legacy MVC
  layout and the observation-type-edit sibling pattern. Carries forward
  the Round-2 wins (inline option-text editing with the index-drift fix,
  modeling-purpose create gate, data-type freeze-after-create, modeling
  attrs description-only). Save redirects to the detail page (matches
  legacy controller behavior).
- Removes the now-unused custom-attribute-type-modal component + dead
  DialogService injections in the grid and detail page.
- Switches the options editor inline inputs to form-field for house-style
  consistency.

Detail page parity:
- Replaces the bullet list of BMP Type names with a table showing
  per-BMP-Type counts (Number of Observation Types, Number of Treatment
  BMPs), mirroring the legacy detail grid. Backed by a new
  TreatmentBMPTypeUsageDto projected via TreatmentBMPType nav properties.

Delete cascade fix:
- CustomAttributeTypes.DeleteAsync was firing entity.DeleteFull(dbContext)
  without awaiting it. Because DeleteFull's body is a chain of
  ExecuteDeleteAsync calls, the controller returned 204 NoContent while
  the cascade raced against DbContext disposal — explains JAR's 5/11/26
  retest finding 'said it deleted but it's still there'. One-line await
  fix; deletion now actually commits.
- Delete error toasts now surface the backend's ProblemDetails message
  instead of a generic 'an error occurred', and the ConfirmService Promise
  chain catches dismissals so X/Escape doesn't reject silently.
- Adds a one-line discoverability note above the grid explaining that
  Modeling attributes are system-managed and intentionally lack a Delete
  affordance.
