+++ Neptune.Web/src/app/pages/manage/custom-attribute-type-edit/custom-attribute-type-edit.component.scss
@@ -1,0 +1,34 @@
+@use "/src/scss/abstracts" as *;
+
+.page-footer {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    gap: $spacing-300;
+    margin-top: $spacing-500;
+    padding-top: $spacing-400;
+    border-top: 1px solid $gray-light;
+}
+
+.options-editor {
+    display: flex;
+    flex-direction: column;
+    gap: $spacing-200;
+
+    &__item,
+    &__add {
+        display: flex;
+        align-items: center;
+        gap: $spacing-200;
+    }
+
+    &__text {
+        flex: 1;
+    }
+
+    // form-field is inline by default; needs block + flex sizing to fit in the row.
+    &__field {
+        display: block;
+        flex: 1;
+    }
+}
