--- Neptune.Web/src/app/pages/manage/custom-attribute-type-modal/custom-attribute-type-modal.component.html
@@ -1,94 +1,0 @@
-<div class="modal">
-<div class="modal-header">
-    <h3>{{ isEdit ? "Edit" : "Add" }} Custom Attribute Type</h3>
-</div>
-<div class="modal-body">
-    <app-alert-display></app-alert-display>
-    @if (isModelingAttribute) {
-        <div class="alert alert-info mb-3">
-            This is a modeling attribute. Only the description can be edited.
-        </div>
-    }
-    <form [formGroup]="formGroup">
-        <div class="grid-12">
-            <form-field class="g-col-12" [formControl]="formGroup.controls.CustomAttributeTypeName"
-                fieldLabel="Name" [type]="FormFieldType.Text" [required]="true" placeholder="Attribute Name">
-            </form-field>
-            <form-field class="g-col-6" [formControl]="formGroup.controls.CustomAttributeDataTypeID"
-                fieldLabel="Data Type" [type]="FormFieldType.Select" [required]="true"
-                placeholder="Select Data Type" [formInputOptions]="dataTypeOptions">
-            </form-field>
-            <form-field class="g-col-6" [formControl]="formGroup.controls.CustomAttributeTypePurposeID"
-                fieldLabel="Purpose" [type]="FormFieldType.Select" [required]="true"
-                placeholder="Select Purpose" [formInputOptions]="purposeOptions">
-            </form-field>
-            <form-field class="g-col-6" [formControl]="formGroup.controls.MeasurementUnitTypeID"
-                fieldLabel="Measurement Unit" [type]="FormFieldType.Select"
-                placeholder="Select Unit (optional)" [formInputOptions]="unitOptions">
-            </form-field>
-            <form-field class="g-col-6" [formControl]="formGroup.controls.IsRequired"
-                fieldLabel="Is Required" [type]="FormFieldType.Select"
-                [formInputOptions]="booleanOptions">
-            </form-field>
-            <form-field class="g-col-12" [formControl]="formGroup.controls.CustomAttributeTypeDescription"
-                fieldLabel="Description" [type]="FormFieldType.Textarea" placeholder="Description (optional)">
-            </form-field>
-            <form-field class="g-col-12" [formControl]="formGroup.controls.CustomAttributeTypeDefaultValue"
-                fieldLabel="Default Value" [type]="FormFieldType.Text" placeholder="Default value (optional)">
-            </form-field>
-        </div>
-
-        @if (showOptionsEditor) {
-            <h4 class="mt-3">Options</h4>
-            <p class="system-text">Add the available options for this pick-list or multi-select attribute. Click an option to edit its text.</p>
-            <div class="options-editor">
-                @for (option of optionsList(); track $index) {
-                    <div class="options-editor__item">
-                        @if (editingOptionIndex() === $index) {
-                            <!-- NPT-1038 rework: inline-edit affordance. Replaces the prior delete-and-re-add
-                                 dance for fixing typos. Enter saves, Escape cancels. -->
-                            <input #editInput type="text" class="form-control form-control-sm options-editor__edit-input"
-                                [value]="option"
-                                (keydown.enter)="saveOptionEdit($index, editInput.value); $event.preventDefault()"
-                                (keydown.escape)="cancelOptionEdit(); $event.preventDefault()">
-                            <button type="button" class="btn btn-sm btn-primary" (click)="saveOptionEdit($index, editInput.value)" title="Save">
-                                <i class="fa fa-check"></i>
-                            </button>
-                            <button type="button" class="btn btn-sm btn-secondary-outline" (click)="cancelOptionEdit()" title="Cancel">
-                                <i class="fa fa-times"></i>
-                            </button>
-                        } @else {
-                            <!-- Plain text + a separate Edit button. The pencil <button> is the
-                                 canonical edit trigger so keyboard / screen-reader users get the
-                                 same affordance as mouse users — no <span (click)> shortcut. -->
-                            <span class="options-editor__text">{{ option }}</span>
-                            <button type="button" class="btn btn-sm btn-secondary-outline" (click)="startEditOption($index)" [attr.aria-label]="'Edit option ' + option" title="Edit">
-                                <i class="fa fa-pencil"></i>
-                            </button>
-                            <button type="button" class="btn btn-sm btn-danger-outline" (click)="removeOption($index)" [attr.aria-label]="'Delete option ' + option" title="Delete">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        }
-                    </div>
-                }
-                <div class="options-editor__add">
-                    <input type="text" class="form-control form-control-sm" placeholder="New option..."
-                        [value]="newOptionText()" (input)="newOptionText.set($any($event.target).value)"
-                        (keydown.enter)="addOption(); $event.preventDefault()">
-                    <button type="button" class="btn btn-sm btn-primary-outline" (click)="addOption()">
-                        <i class="fa fa-plus"></i> Add
-                    </button>
-                </div>
-            </div>
-        }
-
-        <!-- NPT-1038 rework: "Used by Treatment BMP Types" moved out of this modal
-             to the dedicated /manage/custom-attributes/:id detail page so the modal
-             stays focused on edit. -->
-    </form>
-</div>
-<div class="modal-footer">
-    <button class="btn btn-primary" (click)="save()" [disabled]="formGroup.invalid">Save</button>
-    <button class="btn btn-secondary-outline" (click)="cancel()">Cancel</button>
-</div>
-</div>
