+++ Neptune.Web/src/app/pages/manage/custom-attributes.component.ts
@@ -3,7 +3,6 @@ import { AsyncPipe } from "@angular/common";
 import { Router } from "@angular/router";
 import { ColDef } from "ag-grid-community";
 import { BehaviorSubject, Observable, shareReplay, switchMap } from "rxjs";
-import { DialogService } from "@ngneat/dialog";
 import { PageHeaderComponent } from "src/app/shared/components/page-header/page-header.component";
 import { AlertDisplayComponent } from "src/app/shared/components/alert-display/alert-display.component";
 import { NeptuneGridComponent } from "src/app/shared/components/neptune-grid/neptune-grid.component";
@@ -16,7 +15,6 @@ import { CustomAttributeTypeService } from "src/app/shared/generated/api/custom-
 import { CustomAttributeTypeDto } from "src/app/shared/generated/model/custom-attribute-type-dto";
 import { CustomAttributeTypePurposeEnum } from "src/app/shared/generated/enum/custom-attribute-type-purpose-enum";
 import { NeptunePageTypeEnum } from "src/app/shared/generated/enum/neptune-page-type-enum";
-import { CustomAttributeTypeModalComponent } from "src/app/pages/manage/custom-attribute-type-modal/custom-attribute-type-modal.component";
 
 @Component({
     selector: "custom-attributes",
@@ -25,7 +23,7 @@ import { CustomAttributeTypeModalComponent } from "src/app/pages/manage/custom-a
     template: `
         <page-header pageTitle="Custom Attribute Types" [templateRight]="addButton" [customRichTextTypeID]="NeptunePageTypeEnum.ManageCustomAttributeTypesList"></page-header>
         <ng-template #addButton>
-            <button class="btn btn-primary" (click)="openAddModal()">
+            <button class="btn btn-primary" (click)="openAdd()">
                 <i class="fa fa-plus"></i> Add Custom Attribute Type
             </button>
         </ng-template>
@@ -33,6 +31,10 @@ import { CustomAttributeTypeModalComponent } from "src/app/pages/manage/custom-a
         <app-alert-display></app-alert-display>
 
         <div class="page-body">
+            <p class="system-text mb-2">
+                <i class="fa fa-info-circle"></i>
+                Modeling attributes are system-managed and cannot be deleted. The Delete action only appears for Other Design and Maintenance attributes.
+            </p>
             @if (customAttributeTypes$ | async; as data) {
                 <neptune-grid
                     [rowData]="data"
@@ -48,7 +50,6 @@ import { CustomAttributeTypeModalComponent } from "src/app/pages/manage/custom-a
 export class CustomAttributesComponent implements OnInit {
     private customAttributeTypeService = inject(CustomAttributeTypeService);
     private utilityFunctionsService = inject(UtilityFunctionsService);
-    private dialogService = inject(DialogService);
     private alertService = inject(AlertService);
     private confirmService = inject(ConfirmService);
     private router = inject(Router);
@@ -86,7 +87,7 @@ export class CustomAttributesComponent implements OnInit {
                     {
                         ActionName: "Edit",
                         ActionIcon: "fas fa-edit",
-                        ActionHandler: () => this.openEditModal(row),
+                        ActionHandler: () => this.router.navigate(["/manage/custom-attributes", id, "edit"]),
                     },
                 ];
                 if (!isModeling) {
@@ -114,30 +115,8 @@ export class CustomAttributesComponent implements OnInit {
         ];
     }
 
-    openAddModal(): void {
-        const dialogRef = this.dialogService.open(CustomAttributeTypeModalComponent, {
-            data: { mode: "add" },
-            width: "700px",
-        });
-        dialogRef.afterClosed$.subscribe((result) => {
-            if (result) {
-                this.alertService.pushAlert(new Alert("Custom attribute type created.", AlertContext.Success));
-                this.reload$.next();
-            }
-        });
-    }
-
-    private openEditModal(row: CustomAttributeTypeDto): void {
-        const dialogRef = this.dialogService.open(CustomAttributeTypeModalComponent, {
-            data: { mode: "edit", customAttributeType: row },
-            width: "700px",
-        });
-        dialogRef.afterClosed$.subscribe((result) => {
-            if (result) {
-                this.alertService.pushAlert(new Alert("Custom attribute type updated.", AlertContext.Success));
-                this.reload$.next();
-            }
-        });
+    openAdd(): void {
+        this.router.navigate(["/manage/custom-attributes/new"]);
     }
 
     private confirmDelete(row: CustomAttributeTypeDto): void {
@@ -159,11 +138,17 @@ export class CustomAttributesComponent implements OnInit {
                         this.alertService.pushAlert(new Alert("Custom attribute type deleted.", AlertContext.Success));
                         this.reload$.next();
                     },
-                    error: () => {
-                        this.alertService.pushAlert(new Alert("An error occurred while deleting the custom attribute type.", AlertContext.Danger));
+                    error: (err) => {
+                        // Surface the backend's ProblemDetails message when available so PO/dev
+                        // can diagnose a failed delete (FK conflict, permission, etc.) instead
+                        // of seeing the generic toast we used to show.
+                        const detail = err?.error?.detail ?? err?.error?.title
+                            ?? "An error occurred while deleting the custom attribute type.";
+                        this.alertService.pushAlert(new Alert(detail, AlertContext.Danger));
                     },
                 });
-            });
+            })
+            .catch(() => { /* dismissed via X/Escape — no-op */ });
     }
 
     // ConfirmModalComponent renders message via [innerHtml] + bypassSecurityTrustHtml,
